Refactor sticky logic out of PageControls for reuse
